requirements.txt
# 核心数据处理库
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# 数据库相关
duckdb>=0.8.0

# 聚宽数据接口
jqdatasdk>=1.8.0

# Tushare数据接口
tushare>=1.2.89

# 日期时间处理
python-dateutil>=2.8.0

# 类型提示支持
typing-extensions>=4.0.0

# 数据分析和可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 统计分析
statsmodels>=0.13.0
sklearn>=1.1.0

# 并发处理
concurrent-futures>=3.1.1

# 配置管理
pyyaml>=6.0
toml>=0.10.0

# 命令行界面
click>=8.0.0

# 进度条显示
tqdm>=4.64.0

# 配置文件验证
jsonschema>=4.0.0

# 缓存支持
cachetools>=5.0.0

# 重试机制
retrying>=1.3.3

# 数据验证
pydantic>=1.10.0

# 开发和测试工具（可选）
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=0.991